Make ac-quiz more robust and cleaner
