+++ ac/ac-quiz/src/config.js
@@ -1,0 +1,42 @@
+// @flow
+
+export default {
+  type: 'object',
+  properties: {
+    collab: {
+      type: 'boolean',
+      title: 'Collaborative?'
+    },
+    justify: {
+      type: 'boolean',
+      title: 'Students must justify their answers'
+    },
+    MCQ: {
+      title: 'MCQ',
+      type: 'array',
+      items: {
+        type: 'object',
+        title: 'New Question',
+        properties: {
+          question: {
+            type: 'string',
+            title: 'Question'
+          },
+          answers: {
+            type: 'array',
+            title: 'Possible answers',
+            items: {
+              type: 'object',
+              properties: {
+                answer: {
+                  type: 'string',
+                  title: 'Answer'
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+};
